ESModule conversion
Convert imports from `ChromeUtils.import` to
`ChromeUtils.importESModule`.

This necessitates changing the minimum compatible version of the
add-on to 126.0, since that's when the ES Modules were added.

I'm not going to support backward-compatibiliity here. People who want
to keep using TB versions prior to 126 can use an old version of the
add-on.